+++ sdk/monitor/azure-monitor-opentelemetry-exporter/CHANGELOG.md
@@ -10,6 +10,8 @@
     ([#25845](https://github.com/Azure/azure-sdk-for-python/pull/25845))
 - Implement attach statsbeat
     ([#25956](https://github.com/Azure/azure-sdk-for-python/pull/25956))
+- Implement feature statsbeat
+    ([#25956](https://github.com/Azure/azure-sdk-for-python/pull/25956))
 
 ### Breaking Changes